refactor(doctor): split workspace check modules
