+++ .sampo/changesets/923-split-doctor-modules.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split workspace block and feature doctor checks into focused category modules while preserving doctor output order.
